backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-docx
python-pptx
genanki
pydantic
langchain
langchain-community
langchain-openai
langchain-classic
langgraph
langsmith
python-dotenv
chromadb
langchain-chroma
tenacity
python-jose[cryptography]
passlib[bcrypt]
email-validator
reportlab
